Fatsecret token

+++ app/src/main/java/com/example/btl_android_project/entity/RecipeIngredient.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 data class RecipeIngredient(
     val id: Int = 0,
     val recipeId: Int = 0,
-    val quantity: Float = 0f,
+    val quantity: Double = 0.0,
     val unit: String = ""
 ) : Serializable
